requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.4.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.40.37
botocore==1.40.37
Brotli==1.1.0
cachetools==5.5.0
captcha==0.6.0
Cerberus==1.3.7
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.3
chromadb==0.4.13
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
cssselect2==0.8.0
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.14
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
Django==5.2.1
django-cors-headers==4.7.0
django-environ==0.10.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docx2txt==0.8
exceptiongroup==1.2.2
extra-streamlit-components==0.1.71
fastapi==0.115.12
feedparser==6.0.12
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.2
Genshi==0.7.9
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.2.0
grpcio==1.66.1
grpcio-tools==1.66.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
h5py==3.13.0
hpack==4.0.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.4
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.54
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langsmith==0.3.32
llama-cloud==0.0.17
llama-index==0.11.11
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.11
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.3.1
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.2.9
llama-index-multi-modal-llms-openai==0.2.1
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.6
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
mdurl==0.1.2
mem0ai==0.1.11
monotonic==1.6
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
narwhals==1.8.2
neo4j==5.24.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==2.2.5
onnx==1.17.0
onnxruntime==1.21.1
onnxruntime-training==1.15.1
openai==1.75.0
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
pillow==11.2.1
platformdirs==4.2.2
polib==1.2.0
portalocker==2.10.1
posthog==3.25.0
propcache==0.3.1
protobuf==6.30.2
psycopg2-binary==2.9.10
pulsar-client==3.6.1
pwdlib==0.2.1
pyarrow==17.0.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydeck==0.9.1
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.9.0
PyMuPDF==1.26.4
pypdf==5.4.0
PyPDF2==3.0.1
pyphen==0.17.2
PyPika==0.48.9
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-magic==0.4.27
python-sql==1.6.0
python-stdnum==2.1
pytz==2024.2
pywin32==306
PyYAML==6.0.2
qdrant-client==1.11.1
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.11.6
relatorio==0.11.1
reportlab==4.4.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.20.0
s3transfer==0.14.0
sgmllib3k==1.0.0
shellingham==1.5.4
simpleeval==1.0.3
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.10.4
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.46.2
streamlit==1.38.0
streamlit-authenticator==0.3.3
striprtf==0.0.26
sympy==1.13.3
tenacity==9.1.2
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.1
toml==0.10.2
tornado==6.4.1
tqdm==4.67.1
trytond==7.6.1
trytond_country==7.6.0
trytond_party==7.6.0
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
virtualenv==20.26.3
watchdog==4.0.2
watchfiles==1.0.5
weasyprint==65.1
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.16.0
yarl==1.20.0
zopfli==0.2.3.post1
zstandard==0.23.0
